+++ client/src/pages/Login.js
@@ -3,53 +3,55 @@ import {
     makeStyles,
     Button
 } from '@material-ui/core';
-import {Link} from "react-router-dom";
+import {Link} from 'react-router-dom';
 
-import {CustomOutlinedInput} from "../components/inputs";
-import CenteredPageHeader from '../components/centered-page-header';
+import {CustomOutlinedInput} from '../components/Inputs';
+import CenteredPageHeader from '../components/CenteredPageHeader';
 
 const useStyles = makeStyles({
     pageContent: {
+        // Alignment
         display: 'flex',
         flexDirection: 'column',
         alignItems: 'center',
+        //Spacing & Page width
         margin: '0 auto',
         padding: '40px 30px',
         maxWidth: 500
     },
     form: {
-        width: '100%',
+        // Alignment
         display: 'flex',
         flexDirection: 'column',
-        alignItems: 'center'
+        alignItems: 'center',
+        // Form sizing
+        width: '100%',
     },
     lastInput: {
+        // Last input has spacing to separate from bottom
+        // according to specs.
         marginBottom: 80
     }
 });
 
-const SignUp = () => {
+const Login = () => {
     const classes = useStyles();
-    const [nameValue, setName] = useState('');
+
+    // Controlled form paramaters
     const [emailValue, setEmail] = useState('');
     const [passwordValue, setPassword] = useState('');
+    const [confirmPassword, setConfirmPassword] = useState('');
 
     return (
         <main className={classes.pageContent}>
+            {/* Page Header */}
             <CenteredPageHeader
-                title="Create an Account"
-                descriptionText={(
-                    <span>Already a member? <Link to="/login">Login</Link></span>
-                )}
+                title="Member Login"
+                descriptionText={<span>New here? <Link to="/signup">Sign Up</Link></span>}
             />
-            <form className={classes.form}>
-                <CustomOutlinedInput
-                    name="name"
-                    value={nameValue}
-                    label="Name"
-                    onChange={setName}
-                    required
-                />
+
+            {/* Login Form */}
+            <form className={classes.form} noValidate>
                 <CustomOutlinedInput
                     name="email"
                     value={emailValue}
@@ -65,13 +67,20 @@ const SignUp = () => {
                     onChange={setPassword}
                     type="password"
                     required
+                />
+                <CustomOutlinedInput
+                    name="confirmPassword"
+                    value={confirmPassword}
+                    label="Confirm Password"
+                    onChange={setConfirmPassword}
+                    type="password"
+                    required
                     className={classes.lastInput}
                 />
-
-                <Button variant="contained" color="primary">Create account</Button>
+                <Button variant="contained" color="primary">Login</Button>
             </form>
         </main>
-    )
+    );
 };
 
-export default SignUp
+export default Login;
